refactor(mistral3.5): rename module mistral3_vlm_fp8 → mistral3_vlm
Shorten the directory name; the "_fp8" suffix was redundant with the
class's own Mistral3FP8* prefix. Update imports in the module itself,
the resolver hook attribute, the TP-plan qualname registration, and the
side-effect import in recipes/vlm/finetune.py.
